Docstrings

Docstrings

Setfield.Lens — Type.

Lens

A Lens allows to access or replace deeply nested parts of complicated objects.

Example

julia> using Setfield

julia> struct T;a;b; end

julia> t = T("AA", "BB")
T("AA", "BB")

julia> l = @lens _.a
(@lens _.a)

julia> get(l, t)
"AA"

julia> set(l, t, 2)
T(2, "BB")

julia> t
T("AA", "BB")

julia> modify(lowercase, l, t)
T("aa", "BB")

Interface

Concrete subtypes of Lens have to implement

set(lens, obj, val)

get(lens, obj)

These must be pure functions, that satisfy the three lens laws:

get(lens, set(lens, obj, val)) == val (You get what you set.)

set(lens, obj, get(lens, obj)) == obj (Setting what was already there changes nothing.)

set(lens, set(lens, obj, val1), val2) == set(lens, obj, val2) (The last set wins.)

See also @lens, set, get, modify.

source

Base.get — Function.

get(l::Lens, obj)

Access a deeply nested part of obj. See also Lens.

source

Setfield.modify — Function.

modify(f, l::Lens, obj)

Replace a deeply nested part x of obj by f(x). See also Lens.

source

Setfield.set — Function.

set(l::Lens, obj, val)

Replace a deeply nested part of obj by val. See also Lens.

source

Setfield.@lens — Macro.

@lens

Construct a lens from a field access.

Example

julia> using Setfield

julia> struct T;a;b;end

julia> t = T("A1", T(T("A3", "B3"), "B2"))
T("A1", T(T("A3", "B3"), "B2"))

julia> l = @lens _.b.a.b
(@lens _.b.a.b)

julia> get(l, t)
"B3"

julia> set(l, t, 100)
T("A1", T(T("A3", 100), "B2"))

julia> t = ("one", "two")
("one", "two")

julia> set((@lens _[1]), t, "1")
("1", "two")

source

Setfield.@set — Macro.

@set assignment

Return a modified copy of deeply nested objects.

Example

julia> using Setfield

julia> struct T;a;b end

julia> t = T(1,2)
T(1, 2)

julia> @set t.a = 5
T(5, 2)

julia> t
T(1, 2)

julia> t = @set t.a = T(2,2)
T(T(2, 2), 2)

julia> @set t.a.b = 3
T(T(2, 3), 2)

source